rvlan: generate RADIUS VLAN options
The RADIUS VLAN options can now be generated with the default values, which
are set in dot-config.json. If non-default values should be set, then they
have to be given in switches.json with 'rvlan' object. An example below
shows that RADIUS VLAN is enabled and the wri1 port is not authenticated:
"rvlan" : {
    "enable" : "y",
    "unauthPorts" : "1"
}