config: set pseudo parameters for the copper SFPs
Copper SFPs can be used for non-WR ports and are listed in SFP database.
But regarding the missing parameters of the copper SFPs, an error message
is generated: Config item "SFP_PARAMS": parameter "wl_txrx" is wrong.
Trick to suppress such error message is to use the typical parameters of
the fiber SFP for the copper SFPs.