localmaster: add a new layer to allow control traffic between 2 Datamasters
In order to test interaction of Datamasters in the Integration network
there are installed 2 Datamasters for the SIS100 and Unilac machines.
Each Datamaster is connected to its own Localmaster WRS:
'integration_localmaster' and 'uni_int_localmaster'.
'ctrl_sync' is access port dedicated for DMs, 'sync' is trunk port between
Localmasters.